Fix default autoComplete in Input component; move type info into `types.ts`
